BUG: Cannot properly export a class. Need to work on JS Classes
